Implements the encryption side of the AES-GCM mode of operation.
Creates a proc to process plaintext and an additional authenticated data (AAD) to produce ciphertext and an authentication tag, as described in NIST 800-38D ("Recommendation for Block Cipher Modes of Operation: Galois/Counter Mode (GCM) and GMAC").

Currently tested with a few handwritten test cases (with results compared to the BoringSSL implementation) - a broader random "sweep" test will come subsequently.

PiperOrigin-RevId: 471098365