This fixes an issue observed during internal development. When attempting to convert a function to IR, an error was encountered: "Not enough symbolic bindings to convert function: umax; need {N} got {}". The code in question was referencing am imported [top-level] constant that was defined via a series of std::umax invocations.
Ultimately, the root cause of the error was that type deduction and IR conversion order extraction considered constants differently. Type deduction saw ConstRefs as references to elements with precomputed values, i.e., that did not need further investigation. Thus, when typechecking, we didn't record a type info for evaluation of parametric functions what were part of defining a constant value. In IR conversion order extraction, the opposite approach was taken: _every_ function in _every_ path was selected for IR conversion, even when it was part of constant evaluation.

Thus, when we attempted to add such a function to the conversion order, we attempted to verify the presence of all needed parametric values, but since we elided such in type deduction, there were no such parametrics, so the above error was raised.

This fix is to fix "agreement" on constant value determination, and so ir_converter.cc and extract_conversion_order.cc were both updated to refer to precomputed constant values instead of trying to IR-convert their derivation pathway. This has the added bonus of reducing emission of unnecessary IR.

PiperOrigin-RevId: 472817903