Make pattern matching a clamp work on selects when number of cases isn't 2.
Previously, matching on a sel() to check for clamping assumed the select had two cases (and would crash if it had fewer). It seems that 2-case sels can be rewritten by passes to have one case and a default.

This change checks how many cases the select has (including a default value) before trying to pattern match a clamp, and handles the default value if it is present.

PiperOrigin-RevId: 475660417