Clean up symbolic bindings for function converter.
`GetSymbolicBindingsTuple()` excludes module-level constants from symbolic bindings and is otherwise the same as `GetSymbolicBindings()`. Besides being a bit of a misnomer (perhaps from the time when Python was ported to C++), this seems to be an anti-feature. However, `GetSymbolicBindingsTuple()` is used instead of `GetSymbolicBindings()`. This CL replaces the `GetSymbolicBindingsTuple()` usages with `GetSymbolicBindings()`, as well as some other small cleanup. It also adds a test with a function parameter shadowing a module-level constant that was broken before.

PiperOrigin-RevId: 476479764