[delay_model] Add decorating delay model.
This allows to put an interposing layer that will be able to adjust the delay
model estimates on a by-node-id basis in the scheduling process.

This will be used to implement a facility that lets us say "insert a N ps
overhang at the start/end of cycle P${X}" -- the infrastructure will be able to
see what nodes fall on the boundary of the scheduled pipeline at the start of
P${X}, increase they delay by the requested amount, and re-schedule.

PiperOrigin-RevId: 477557418