[codegen] Refactor codegen flags to proto populator.
This will allow us to a) reuse the flags more easily from the upcoming delay
munger (which needs to invoke the codegen_main facilities like a subroutine)
and b) read the now-numerous flags from a proto to save on manual configuration
specification via flags (e.g. in programmatic cases or ones with a large number
of flag modifications from the defaults).

PiperOrigin-RevId: 477801874