Various optimizations to the JIT. First, deduplicate operand arguments to node functions. Previously, if a node X appeared multiple times as an operand to node Y, then X would appear multiple times as an argument to the JITted function implementing Y. With this change, X will only be passed in once.
Second, optimize the packing function. Previously, a potentially enormously wide integer value is created and sequentially shifted to unpack it. Now the packed value is left in memory, and unpacking is done by indexing into the packed buffer to extract the values.

Third, don't build a packed wrapper for JITted proc functions as it was unused.

PiperOrigin-RevId: 478653509