Adding proc-init-by-global-constant support to the fuzzer.
We're shortly moving to specifing initial proc state via a proc-scoped function ("init()"), so this is a necessary precondition for that. The new ValueGenerator::GenerateConstantDef() routine isn't comprehensive - there are input TypeAnnotations that aren't yet supported, but all types produced by the fuzzer _are_ currently supported. If the set generated by AstGenerator expands, then there's a chance that ValueGenerator will need to be updated, as well.

PiperOrigin-RevId: 485370081