DSLX: Fixing crash and adding parser error when referencing proc members inside a config function.
It doesn't make sense to reference proc members inside a proc config function, as they don't have defined values until after the function completes. Doing so led to a crash in IR conversion for that reason. I added a safety check to ProcConfigIrConverter to avoid that, and added a check (and useful error message!) during parsing to catch this sort of error.

Fixes issue #770.

PiperOrigin-RevId: 485949197