DSLX: Fixing parsing to enable creation of arrays of ColonRef-to-type types.
Previously, the code was assuming that a ColonRef followed by a '[' could only be an index op (as if the ColonRef pointed to a constant array), but the construct `import_module::Type[2]:[...]` is (or should be) a valid array declaration, so the parsing was updated to recognize this possibility.

As an aside, the function `Parser::ParseTerm()` is becoming quite complex and has some hard-to-infer assumptions along its path (such as assuming the constructs entering the second half of the function must be of some particular types), so it'd benefit from a refactoring in the near future.

Part of resolving issue #697.

PiperOrigin-RevId: 486230196