Add bounds checking for slice index values.
Index values (e.g., the `i7:0` in `a[i7:0 :]`) are typed, and previously we didn't do any early bounds checking on them, which would cause out-of-bounds errors in cases like:

```
let a: u8[128] = ...
a[u2:0 :]
```

Slice indices must be of the same type, so in the above example, the implied end index would also be a u2, but it would need to store the value 127 in it. Late in evaluation, we'd see this problem and explode. This change does earlier bounds checking (during type deduction) so we can catch and gracefully report this error.

PiperOrigin-RevId: 487548585