[FUZZER] Generate Verilog with flop_input_kind equal to zerolatency for samples containing non-blocking receives.
Non-blocking receives introduce non-deterministic behavior (more than one set of possible valid results).
To facilitate the matching of the results in the fuzzer of other stages to the simulation stage, samples with non-blocking receives generate Verilog with `flop_input_kind=zerolatency` to mimic the behavior evaluated in other stages in the fuzzer.

PiperOrigin-RevId: 491707878