[deps] Add a first example use of RapidCheck.
This will allow us to create property based tests in XLS' C++ tests more
easily. Not to be confused with the user-level fuzzing infrastructure -- this
is more intended for use on testing library components written in C++.

RapidCheck relies on exceptions for some features, which is unfortunate. Also,
due to the fact the "checker" combinators are different between gtest and
rapidcheck, it can be difficult to reuse code between property-based tests and
unit tests (e.g. if a property based test produces a stimulus we want to use as
a unit test vector, and we want to reuse the testing/checking functionality).

We may, in the limit, need to replace RapidCheck in the future with something
more Google-convention friendly, but even in that case, the property-based test
specifications will be venerable and will be able to be ported in the event of
a library change.

Fixes google/xls#476

PiperOrigin-RevId: 492000412