[rapidcheck] Add rapidcheck seed to XLS' gunit main binary.
Generally we assume unit tests will behave deterministically which is useful
e.g. for not needing to determine whether flakes are "expected or not". In the
mode we typically work in, flakes are never expected.

By setting the seed up front we use quickcheck for its "better than a unit
test" properties but obtain determinism in a given build. (At least... given
the ordering of the gunit test cases.)

This also increases the number of test cases from the default, since we're
using it deterministically, to get a bit more coverage, and turns on verbose
mode so we can see the test cases and how many are skipped.

See configuration documentation here:
https://github.com/emil-e/rapidcheck/blob/master/doc/configuration.md

PiperOrigin-RevId: 492056551