[DSLX] Destructors in source file for AST nodes. This makes RTTI-based dynamic casting work.
When building the pybind extension with LLVM toolchain we observed that
these casts were returning null, presumably because RTTI data was being
discarded in the shared object we create via `pybind_extension`.

Notably this doesn't get rid of all the errors yet, there are still some
miscomparisons in fuzz tests, but the cast-based crashes are gone.

PiperOrigin-RevId: 494233355