[DSLX] Parametric bindings are resolved in type evaluation.
Previously a parametric constant value could not be used in a subsequent
parametric expression; e.g.

```
fn clamp<BITS: u32, MIN_VALUE: sN[BITS] = sN[BITS]:0>(x: sN[BITS]) -> sN[BITS] {
                                          ^~~~~~~~~^ expression for parametric binding, uses BITS
  ...
}
```

The solution was to create *scoped type information* that is used when
evaluating the parametric part of the signature. This gives us a place to note
that, when instantiated as `clamp<u32:5>`, `sN[BITS]` resolves via the concrete
value of `BITS` to `sN[5]` -- that resolution is now placed in the scoped type
information. This way, when we invoke the interpreter to evaluate the
expression `sN[BITS]:0`, `sN[BITS]` is appropriately resolved to a fully
concrete type.

(Previously, the interpreter would give an error that `sN[BITS]` still had a
parametric dimension `BITS` that had not been resolved.)

There is still work to be done -- the ParametricInstantiator helper is used in
**type inference**, but it does not keep a record of evaluated values from
parametric expressions that can be picked up by **IR conversion**. Follow on
step will be to record these evaluated parametric values in the type
information, e.g. in the value side of the caller-to-callee instantiation map:
https://github.com/google/xls/blob/59f6ae8a0b51c3cd2c73bc869bdd8dc87ff81645/xls/dslx/type_info.h#L54
-- from there it would be naturally picked up in the IR conversion phase's use
of TypeInfo.

Towards #818

PiperOrigin-RevId: 501429211