[dslx] Rename SymbolicBindings to ParametricEnv.
This change is primarily because "symbolic bindings" is not really a term of
art. When we evaluate parametric values, in normal programming language terms,
we're just taking parametric expressions and pushing them through a parametric
environment to get a value (or perhaps a partially-evaluated parametric
expression). We largely keep parametric environments around to know what we
should instantiate parametric functions with. "symbolic bindings", as it was
previously, introduced a new term that's different from "environment", and
"symbolic" doesn't really mean very much, it just indicated that parametric
symbols were free variables that got bound by parametric instantiation.

PiperOrigin-RevId: 502610560