Add --z3_rlimit parameter to xlscc. Creates a path to unroll loops less "strictly", avoiding spending too much time trying to prove earlier break, when a later break is much easier to prove.
For example, in https://github.com/google/xls/issues/854 , explicit loop bounds are set "for (char k = 0; k < 4; k++)", but xlscc spends a long time trying to prove that iterations 2 and 3 might break. That happens because there is an explosion of logic complexity, probably due to the multiple conditional assignments and breaks in the loop body. However, as a human can easily see, proof becomes trivial on iteration 4.

PiperOrigin-RevId: 508168223