Standardize on http_archive() fetching for workspace dependencies.
The workspace dependencies had a mix of http_archive and git
archives for no particular reason.
Using git is much slower than fetching an http archive and
also requires the `git` binary to be present at fetch time.

So standardize on http_archive.

This has the added benefit that we can double-check what
arrives with the sha256 hash.

PiperOrigin-RevId: 508784740