Provide a script that generates a compilation database.
The compile_commands.json file contains information how the compiler
is invoked on the files. It is used for tools such as the clangd
language server or clang-tidy as well as a bunch of other
static analysis tools to provide their features.

Unfortuantely, bazel does not support this right out of the box,
so this is using a project that uses bazel aspects to extract the
relevant information.

Providing a compilation database at least partially addresses #853

PiperOrigin-RevId: 512781357