Provide a script to run clang-tidy.
clang-tidy is pretty slow to run on all our files, so cache
the output that only changed files will have to be reprocessed
in subsequent runs (changes in .clang-tidy or WORKSPACE will trigger
full reprocessing).

For each cc-file, store the output of clang-tidy content-addressed,
so that a minimum of reprocessing is needed: if a file with a
particular content has been processed with a particular clang-tidy,
that content is cached and retrieved quickly.

Also make sure that cc files are reprocessed if any of their headers
change.

Provide a symbolic link 'xls-clang-tidy-current.out' that points to
the last called clang-tidy output file; thus this can be used
as a stable location.

If `CLANG_TIDY` environment variable is set, use its content
as the clang-tidy binary to call, otherwise fall back to search
in system path.

If `CACHE_DIR` is set, then use that as base of where the cache
is located, otherwise use ~/.cache, ${TMPDIR}, or /tmp whatever is
available first.

Issues: #262
PiperOrigin-RevId: 512800838