[DSLX] Add `extern_verilog` attribute support on functions.
This differs from how we identify `#[test]` functions -- those are done with an
AST node at module scope. Because `#[extern_verilog("foo")]` doesn't change the
execution semantics at the DSL level at all, it's really just a note for IR
conversion, we just tag the function with the attribute and remember to observe
it at IR conversion time.

Just parses for now since we need the IR API to be able to annotate
"behavioral" functions for FFI replacement. So really, this annotation is a
no-op for now, but this adds the necessary parse support that can be picked up
in IR conversion.

No documentation update since it doesn't do anything yet.

(Also deletes a few redundant member vars I noticed on TestFunction AST node.)

Fixes google/xls#892

PiperOrigin-RevId: 515682561