Rewrite run-clang-tidy script in C++ for readability.
The run-clang-tidy-cached script provides efficient caching
of checked files and their dependencies (as a full run
takes about 1600 core minutes for xls: caching is vital).
As such, it is bit more compliated than simply invoking
clang-tidy.

The shell script to deal with that was growing beyond
healthy and was assuming certain features in the
various tools involved (sed, awk, xargs, sort...) that
might not be available everywhere. This, and
complicated quoting rules and esoteric scripting
features also made it fragile, hard to read, and maintain.

This is a rewrite in portable c++17. As such, it is
also much easier to maintain by the C++ authors of
XLS.
Tested on Linux and MacOS but would probably also
work on Windows.

File is designed as C++ script, so can be invoked
directly.

PiperOrigin-RevId: 515691749