Fix RAM write response rewriting.
The RAM rewrite pass rewrites RAM usages, currently mapping abstract RAMs to 1RW RAMs. Abstract RAMs have read request+response and write request+response channels. Receiving on a write response channel returns no data, but is useful for marking write completion. The RAM rewrite pass rewrites a receive on the write response channel to be a receive on the 1RW response channel, discarding the data. This interacts poorly with 1RW RAM codegen, which expects receives on the response channel only after read requests (this avoids latching in garbage).

The fix is to replace the predicate on the new receive corresponding to a write response to  be unconditionally false. This way, a receive on the write response channel will not trigger a receive on the 1RW response channel.

PiperOrigin-RevId: 515728942