[DSLX] Introduction of Statement AST node.
Right now it's somewhat degenerate just to get a stepping stone -- blocks are
expected to contain single statements (which was previously the body
expression).

From here we can start parsing multiple statements (i.e. `TypeAlias`) into
blocks and walk over statements to accumulate bindings. Most meaningful step
there is that we'll need to "unfurl" `Let` AST nodes so their body is not
nested inside (as in ML style) but rather a successive series of statements.

PiperOrigin-RevId: 516598468