[DSLX] Place type mismatch data in a structure on the typechecking context.
Instead of forming an error at the point a mismatch was found, we scribble data
"on the wall" of the current type checking context, and then interpolate it
into string form at the entry point to the type checking subsystem.

This allows us to add more structured data and pass it to a diagnostic
subsystem to form more complex and/or refined error messages for user
consumption.

PiperOrigin-RevId: 516956231