Add DifferenceConstraint, allowing direct specification of SDC constraints
Previously there was no way to add an arbitrary SDC constraint to the
scheduler, which is needed for preventing the scheduler from scheduling
sends and receives in the same cycle during proc clumping.

Additionally, this fixes an issue where the proc clumping pass could have
a NodeInCycle constraint passed in without that node pointer being mapped
to the cloned node when each cycle is extracted into a separate proc.

Could also perhaps be useful for #899 down the line.

PiperOrigin-RevId: 527033771