Small code documentation and const improvements.
Set virtual methods that are leafs to final, which improves
documentation of that fact for the human and opens opportunities
for the compiler for devirtualization.

Mark some functions const. Use inline constexpr instead of
static constexpr in a header, as the latter lets the compiler
wonder why a static variable is not used in files including
this header

PiperOrigin-RevId: 528861310