Allow scheduling wrapper pass to reschedule if new nodes are found.
This will be used to add proc inlining, a pass that adds new nodes, as a scheduling pass. Proc inlining will be run as a scheduling pass after the upcoming pass that legalizes multiple channel operations on a single channel by adding an adapter proc to inline the procs added by the pass.

PiperOrigin-RevId: 530408270