Fix race conditions in module testbench. Previously in the module testbench builder signals were both set and sampled on the negative edge of the clock. This could cause race conditions especially if multiple threads are setting/sampling the same signal. This CL changes the testbench to the following behavior:
* Signals are set one time unit *after* the posedge of the clock.

* Signals are sampled one time unit *before* the posedge of the clock.

This required changing the testbench api to introduce a representation for the moment of time when signals are sampled (one time unit before posedge of the clock). Signals are sampled (`Capture`, `ExpectEq`) using this construct while the setting of signals (which occurs one time unit after posedge clock) is still done directly on the thread.

One notable change in behavior is that the `WaitFor...` (and the new `AtEndOfCycleWhen...`) methods now always result in at least one cycle elapsing.

Also improve the readability of the testbenches with additional comments.

PiperOrigin-RevId: 532819219