[DSLX] Refactor parametric dimensions a bit and fix internal errors.
Some background:

* As an artifact of how the system is currently constructed, we `Deduce()` the
  type signature for parametric functions when the parametric environment is not
  fully known/resolved.
* `Deduce()` gives back a type called `ConcreteTypeDim`, but
  that is somewhat unfortunately named, since `ConcreteTypeDims` can actually
  be `ParametricExpression`s.
* When we `Deduce()` a parameterized function signature that we don't have an
  environment for it turns into a `ConcreteType` that can have
  `ParametricExpression`s in its dimensions.
* The `ParametricExpression`s are a tiny little arithmetic subset of the full
  AST set of operators.

The longer term fix here is likely to only instantiate/check the parametric
signatures when parametric bindings are available -- the current implementation
is an artifact of an initial attempt we made to see if there were obvious
conflicts in the parametric type signature before instantiation; e.g. "can you
tell a signature that they wrote a signature that cannot work for any values
before somebody goes to instantiate it". Switching to be
only-instantiation-driven will make things more akin to C++ templates where
they can have impossible nonsense inside so long as they are never
instantiated, and will allow us to clean up the half-way aspects in the above
bullets that constrain parametric expressions.

In the new world ideally you'd pass the type annotation AST into the parametric
instantiator (instead of a `ConcreteType` with parametric expressions inside as
we do today), which would mean pattern-match-style creation of bindings in the
parametric environment would get more complicated.

Fixes google/xls#967

PiperOrigin-RevId: 533599339