Run TokenDependencyPass after ChannelLegalizationPass.
The ChannelLegalizationPass adds an adapter to arbitrate between multiple channel operations on a single channel. The original channel operations are replaced with three operations:
1. Send the predicate to the adapter
2. Perform the original operation, but now on a new channel coming from/going to the adapter.
3. Receive a completion from the adapter indicating the operation has been performed.

The send of the predicate takes the proc's token param as the input token, but in general this may not be correct. For example, a predicate that is data dependent on a receive will require the token from that receive.

Ideally, the legalization pass would perform analysis to choose the correct token (or later passes will not rely on token dependencies reflecting data dependencies). For now, run TokenDependencyPass to clean up the potentially incorrect tokens.

PiperOrigin-RevId: 534211438