[DSLX] Improve type handling in parametrics
Some buildins like `zero!` did not work with
parametric structs and with arrays of structs.

ParseParametrics returns TypeAnnotation only if
token is a buildin type. In all other cases
identifiers are expected to resolve to TypeRef or
constant values.

This commit changes ParseParametrics to allow for
parametric structs and struct arrays types.
This change only extends Parser capabiliteis.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>