Delay model generation: Update Yosys/ABC script to do sizing of std cells. Also, add more robust checking of the results returned from the synthesis server. Fixes google/xls#1002.
Also catches an issue that happens when the operator implementation
gets optimized to a constant (e.g. sub(a,a) or xor(a,a)).
In this case there are no timing paths.  A fatal error is issued.
Note this is different than the case where the operator is optimized
to wire(s) with zero delay, which is supported.

PiperOrigin-RevId: 537067886