Provide a template parser for FFI code generation.
This allows to provide a template for the code generation for
foreign function calls. The {...} template parameters can include
DSLX expressions that are to be evaluated and expanded when emitting
code.

Simple example could be:
```
#[extern_verilog("pop_count_verilog {fn} (.i({x})),.o({return})")]
fn pop_count(x: u32) -> u8 { ... }
```

This adds the template parsing functionality (does not emit yet on code
generation).

The simple std::string to represent the foreign function name in
ForeignFunctionData is now replaced with the more detailed, pre-parsed template.

Simple syntactic errors (such as mismatched parenthesis)
are emitted in the parsing phase (and thus are also immediately
useful in the language server).

PiperOrigin-RevId: 537943042