Make code navigation a bit easier by stating sources of generated files.
Various files are generated from the specification applying a template,
but when coming across such a file it would be good to know the sources.

This adds information in the 'AUTOMATICALLY GENERATED, do not edit'
headers of render_specification_against_template.py stating the sources
of the specification and the template used.

PiperOrigin-RevId: 538834313