[DSLX:bytecode] Expression-statements should not leave results on the stack.
We need to hold on to a expression-statement result in case it's the last
expression in a block and there is no trailing semicolon (because that is the
case where Rust yields the trailing expression's value). However, we were not
clearing (popping) the result between expression-statements.

Additionally, in fixing that, it became clear that `spawn`
expression-statements were special and not producing any value on the stack --
it is changed in this CL to push a unit value so it is not special vs other
expression-statements.

This also adds a `tuple_index` bytecode just to help maintain runtime
invariants; e.g. when an attribute access knows the left hand side is a struct,
we know it should show up at runtime as a tuple, which helps encode known
emission-time information into the bytecodes (and new bytecodes are generally
cheap, plus ours are currently quite fat).

PiperOrigin-RevId: 540021902