Add strictness as a parameter to the channel IR construct.
Previously, there was a global setting for how to treat multiple channel ops occurring on the same channel. This change moves that global setting to a per-channel parameter in the IR. If unspecified, a default value of 'proven_mutually_exclusive' is assumed. This change allows different channels to have different semantics and also allows frontends the opportunity to configure the channels.

For now, XLSCC uses the very permissive 'arbitrary_static_order' strictness as we are actively developing memory support where this is useful. In the future, this should be user-configurable. google/xls#1023 tracks future frontend support for this feature.

PiperOrigin-RevId: 540404118