Make dfe_pass correctly handle cases where different procs are accessing the same channel.
Previously, dfe_pass followed internal channels on send -> recv and recv -> send paths to see what procs were connected to the top proc. It did not consider cases where the top proc and another proc were both sending to the same external channel.

This change replaces this analysis with a simpler analysis that looks for any procs that use the same channels as the top proc (or any other proc reachable by the top proc).

PiperOrigin-RevId: 542075318