Fix proc inlining's logic that saves non-blocking receives.
In cases where a user is a send on a single-value channel or when users are not necessarily active in the same tick as the receive, proc inlining saves the result of the receive to be used in subsequent ticks. For non-blocking receives, this means the valid signal must also be saved. I missed this in my previous fixes to proc inlining and address it here.

PiperOrigin-RevId: 543585470