[DSLX] Switch smulp signature to return unsigned numbers.
Multiplying signed numbers via this primitive results in partial product
outputs, which are not necessarily signed themselves. Once they are added they
can be converted to a signed number. As a result, this is the more correct
signature:

	(sN[N], sN[N]) -> (uN[N], uN[N])

(A followon change should also make the signature be more similar to std::smul
where it looks like: `sN[A], sN[B] -> (uN[A+B], uN[A+B])`).

PiperOrigin-RevId: 543955481