Fix narrowing pass for mulps.
Split multiply ('mulp') operations are free to return arbitrary tuples of type `(bits[N], bits[N])`. While it is true that the output width of the overall multiply can be narrowed if the sum of the input widths are less than N, care needs to be taken with how the split multiply is narrowed.

The previous implementation narrowed the split multiply and then extended the tuple entries before going to the add. This is correct for the naive mulp implementation that returns (0, a*b), but it is not legal to do in general. The new narrowing implementation only narrows mulps if they are immediately followed by a single add (via tuple_index nodes). The narrowing is performed by replacing both the mulp and the add with narrower versions, and sign extension is performed after the add.

PiperOrigin-RevId: 544745688