Improve default mulp implementations.
The previous naive implementations of smulp and umulp returned (0, a*b). This can mask unsound optimizations, for example sign extending the elements will commute with addition for the naive implementation, but will not commute in general.

This changes default mulp implementations to return (OFFSET, a*b - OFFSET) for a fixed value of OFFSET.

This change also adds coverage to ensure the mulp implementations reveal the unsound optimizations discussed above.

PiperOrigin-RevId: 545770925