[DSLX] Fix a few internal errors to properly displayed userspace errors.
One of the issues here was that the type system refines the set of
possibilities to be fewer than there are *before* type inference, but they were
using the same sets of routines. This introduces a refinement for resolving AST
nodes before/after type checking has been performed so the type system has to
handle the cases and the other subsystems don't.

PiperOrigin-RevId: 547268369