Provide a FirstMatchDelayEstimator.
A DelayEstimator that delegates to a sequence of
other DelayEstimators until the first returns a valid result.

This is in preparation to compose DelayEstimators from the
basic PDK model and delay information for foreign function
calls. The former are typically fixed and compiled in, while
we expect the latter to gather their data from other sources.

PiperOrigin-RevId: 547568339