Augment CodegenPassUnit with block conversion metadata.
After block conversion, a lot of metadata is lost and it becomes difficult to determine things like what stage a node belongs to. This change adds previously internal metadata to CodegenPassUnit. Now, block conversion functions return a CodegenPassUnit instead of just the Block.

The metadata is added to the CodegenPassUnit rather than CodegenPassOptions to support future decomposition of block conversion into incremental passes.

This first cut is intended to be minimally invasive. It might be the case that not all of the previously-internal data structures should be exposed or that the organization should change, but to keep the refactoring manageable I'm changing as little as possible. The main change is that pipelined functions now plumb out pipeline_valid signals.

PiperOrigin-RevId: 547617952