Allow the BuilderBase to publish errors.
Adds a new interface function in the BuilderBase, GetError(), that returns
a detailed error fit for printing to a user, or absl::OkStatus() if no error
is pending.

This is for use by utility functions that compose complex BValues from
smaller ones, and wish to return early if a semantic analysis error
has occurred, without invoking e.g. function_builder.Build() outright.

PiperOrigin-RevId: 548546978