[DSLX:type_system] Slight refactoring of parametric instantiation.
The class hierarchy just gets in the way (remnant from the original way it was
done), starting to peel it apart.

Modernizes the comments -- I think this "eagerly try to evaluate default
parametric expressions" routine maybe possible to rip out in the limit, but we
need to clean it up a bit first before we cut the gordian knot.

PiperOrigin-RevId: 549349899