Provide a baseline FFI delay estimator.
Currently, this only returns a constant value provided
in the scheduling options via a flag.

This allows to use FFI with delay_models other than unit, as
it provides a time for the invoke operation.

This will be updated once we have specified a way for the
user to provide delay estimations from a configuration file.

PiperOrigin-RevId: 549401126