[DSLX:bytecode] Add callback for rollover cases from the bytecode interpreter.
"Rollover cases" are defined as those cases that differ from the infinite
precision version of the operation; i.e. showing the distinction between
signed/unsigned ops on the real number line vs in a two's-complement arithmetic
representation that wraps.

In a subsequent CL this will allow us to emit warnings when rollovers occur
during constexpr evaluation, i.e. during typechecking. It potentially also
could be used in determining rollover coverage information for arithmetic
operations.

Using type information in bytecode emission of adds uncovered that the FIFO
depth for channel declarations was not being type inferred, so this fixes that
along the way.

PiperOrigin-RevId: 549754212