Refactor channel legalization adapter.
1. Predicate receives
Previously, the adapter did a blocking receive on every predicate before doing any channel operation. This prevented use cases like `a = mem[addr]; if (f(a)) a += mem[addr + 1];` because the second predicate depended on the result of a previous read.

Now the adapter does a non-blocking receive on every predicate and makes forward progress even if not all predicates are available.

2. Correctly thread tokens
During channel legalization, use the operation's token as the input when sending predicates.

Previously, the proc token param was used on the theory that it was better to send predicates ASAP (later runs of TokenDependencyPass would correct it if that was optimistic). However, sending predicates ungates the adapter to perform the side-effecting op, so sending predicates ASAP can violate the token dependency order if e.g. side-effects A->B but pred(B)->pred(A).

3. Refactor adapter builder
Improve internal data structures for managing adapter state.

4. Add coverage
Add tests that enforce that the adapter correctly sequences operations.

PiperOrigin-RevId: 551253434